requirements.txt
h5py==3.3.0
hmmlearn==0.2.6
matplotlib==3.4.3
numpy==1.20.3
pandas==1.3.4
Pillow==9.0.0
pymol==0.1.0
scikit_learn==1.0.2
scipy==1.7.1
seaborn==0.11.2
tqdm==4.62.3
